Tools_SDK_Chainlit/pyproject.toml
[project]
name = "tools-sdk-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.13",
]
